extracted code to multi-file setup, added cmd line args
